Add missing (long long) cast of variable, (long long) casts for defines fragnum, fsbtodb, cgsblock, and cgdmin.

+++ sbin/fsck_ffs/inode.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: inode.c,v 1.39 2013/11/01 17:36:18 krw Exp $	*/
+/*	$OpenBSD: inode.c,v 1.40 2013/11/03 02:22:07 krw Exp $	*/
 /*	$NetBSD: inode.c,v 1.23 1996/10/11 20:15:47 thorpej Exp $	*/
 
 /*
@@ -239,8 +239,8 @@ chkrange(daddr_t blk, int cnt)
 	if (cnt > sblock.fs_frag ||
 	    fragnum(&sblock, blk) + cnt > sblock.fs_frag) {
 		if (debug)
-			printf("bad size: blk %lld, offset %lld, size %d\n",
-			    (long long)blk, fragnum(&sblock, blk),
+			printf("bad size: blk %lld, offset %lld, size %lld\n",
+			    (long long)blk, (long long)fragnum(&sblock, blk),
 			    cnt);
 		return (1);
 	}
@@ -249,10 +249,11 @@ chkrange(daddr_t blk, int cnt)
 		if ((blk + cnt) > cgsblock(&sblock, c)) {
 			if (debug) {
 				printf("blk %lld < cgdmin %lld;",
-				    (long long)blk, cgdmin(&sblock, c));
+				    (long long)blk,
+				    (long long)cgdmin(&sblock, c));
 				printf(" blk + cnt %lld > cgsbase %lld\n",
 				    (long long)(blk + cnt),
-				    cgsblock(&sblock, c));
+				    (long long)cgsblock(&sblock, c));
 			}
 			return (1);
 		}
@@ -260,7 +261,8 @@ chkrange(daddr_t blk, int cnt)
 		if ((blk + cnt) > cgbase(&sblock, c+1)) {
 			if (debug)  {
 				printf("blk %lld >= cgdmin %lld;",
-				    (long long)blk, cgdmin(&sblock, c));
+				    (long long)blk,
+				    (long long)cgdmin(&sblock, c));
 				printf(" blk + cnt %lld > sblock.fs_fpg %d\n",
 				    (long long)(blk+cnt), sblock.fs_fpg);
 			}
